finish initial OAP injector program (not tested on vehicle)
